+++ app/admin/clients.rb
@@ -27,6 +27,9 @@ ActiveAdmin.register Client do
     panel "Objets" do
       table_for resource.assets do
         column 'Nom', :name
+        column :description do |asset|
+          simple_format asset.description
+        end
         column "Date d'expiration", :expiration_date do |asset|
           l asset.expiration_date, format: :long
         end